requirements.txt
joblib==1.2.0
numpy==1.24.3
scikit-learn==1.3.0
grpcio==1.67.0
grpcio-tools==1.67.0
google-api-python-client==2.92.0
grpcio-reflection==1.67.0
python-dotenv==1.0.0
numpy==1.24.3
pydantic==1.10.7